# go-ecommerce-api/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "A RESTful E-commerce API built with Go and Gin",
        "title": "E-commerce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/admin/products": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product (Admin only). Valid categories: Electronics, Accessories, Home, Office",
                "tags": [
                    "admin"
                ],
                "summary": "Create a new product",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.Product"
                            }
                        }
                    },
                    "description": "Product data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/products/bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create multiple products at once (Admin only). Valid categories: Electronics, Accessories, Home, Office",
                "tags": [
                    "admin"
                ],
                "summary": "Create multiple products",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.BulkCreateProductsRequest"
                            }
                        }
                    },
                    "description": "Array of products to create",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ProductsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/products/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing product (Admin only)",
                "tags": [
                    "admin"
                ],
                "summary": "Update a product",
                "parameters": [
                    {
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.Product"
                            }
                        }
                    },
                    "description": "Updated product data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a product by ID (Admin only)",
                "tags": [
                    "admin"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "description": "Retrieve all products",
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ProductsResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/category/{category}": {
            "get": {
                "description": "Retrieve products filtered by category",
                "tags": [
                    "products"
                ],
                "summary": "Get products by category",
                "parameters": [
                    {
                        "description": "Product Category",
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ProductsResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/featured": {
            "get": {
                "description": "Retrieve all featured products",
                "tags": [
                    "products"
                ],
                "summary": "Get featured products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ProductsResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/search": {
            "get": {
                "description": "Search products by name or description",
                "tags": [
                    "products"
                ],
                "summary": "Search products",
                "parameters": [
                    {
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ProductsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Retrieve a specific product by its ID",
                "tags": [
                    "products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "models.BulkCreateProductsRequest": {
                "type": "object"
            },
...
            "models.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "models.Product": {
                "type": "object",
                "properties": {
                    "badge": {
                        "type": "string"
                    },
                    "badge_color": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "featured": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "original_price": {
                        "type": "number"
                    },
                    "price": {
                        "type": "number"
                    },
                    "rating": {
                        "type": "number"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "models.ProductResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/models.Product"
                    }
                }
            },
            "models.ProductsResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/models.Product"
                        }
                    }
                }
            },
...
            "models.SuccessResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "message": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}